Updating example to use Evalidator

+++ ExampleScripts/FIAPlotExample.R
@@ -17,33 +17,62 @@ state <- "TN"
 clipSize <- 1610
 showMaps <- TRUE
 
-# get FIADB for state
-FIADBFile <- paste0("FIADB_", state, ".db")
+useEvalidator <- TRUE
+if (useEvalidator) {
+  # code from Jim Ellenwood to read PLOT data from Evalidator
+  # works as of 9/30/2021 but the API may go away in the future
+  library(httr)
 
-# check to see if file exists
-if (!file.exists(FIADBFile)) {
-  fetchFile(paste0("https://apps.fs.usda.gov/fia/datamart/Databases/SQLite_FIADB_", state, ".zip"), paste0(state, ".zip"))
+  # get a table of state numeric codes to use with Evalidator
+  form <- list('colList'= 'STATECD,STATE_ABBR', 'tableName' = 'REF_RESEARCH_STATION',
+                'whereStr' = '1=1',
+                'outputFormat' = 'JSON')
+  urlStr <- 'https://apps.fs.usda.gov/Evalidator/rest/Evalidator/refTablePost?'
+  response <- POST(url = urlStr, body = form, encode = "form") #, verbose())
+  states <- as.data.frame(fromJSON(content(response, type="text")))
+  states<-rename_with(states, ~ gsub("FIADB_SQL_Output.record.", "", .x,fixed=TRUE), starts_with("FIADB"))
+  statecd<-states[states$"STATE_ABBR" == state,"STATECD"]
 
-  # unzip the file
-  unzip(zipfile = paste0(state, ".zip"), exdir = getwd())
+  form = list('colList' = 'CN,PREV_PLT_CN,STATECD,INVYR,MEASYEAR,MEASMON,MEASDAY,LAT,LON,DESIGNCD,KINDCD,PLOT_STATUS_CD',
+               'tableName' = 'PLOT', 'whereStr' = paste0('STATECD=', statecd),
+               'outputFormat' = 'JSON')
+  urlStr = 'https://apps.fs.usda.gov/Evalidator/rest/Evalidator/refTablePost?'
+  response <- POST(url = urlStr, body = form, encode = "form") #, verbose())
 
-  if (file.exists(FIADBFile)) {
-    file.remove(paste0(state, ".zip"))
-  } else {
-    # report a problem unzipping the file
-    cat("*****Could not unzip file: ", paste0(state, ".zip"), "\n")
+  library(jsonlite)
+  PLOT <- as.data.frame(fromJSON(content(response, type = "text")))
+
+  library(dplyr)
+  PLOT <- rename_with(PLOT, ~ gsub("FIADB_SQL_Output.record.","",.x,fixed=TRUE), starts_with("FIADB"))
+} else {
+  # get FIADB for state
+  FIADBFile <- paste0("FIADB_", state, ".db")
+
+  # check to see if file exists
+  if (!file.exists(FIADBFile)) {
+    fetchFile(paste0("https://apps.fs.usda.gov/fia/datamart/Databases/SQLite_FIADB_", state, ".zip"), paste0(state, ".zip"))
+
+    # unzip the file
+    unzip(zipfile = paste0(state, ".zip"), exdir = getwd())
+
+    if (file.exists(FIADBFile)) {
+      file.remove(paste0(state, ".zip"))
+    } else {
+      # report a problem unzipping the file
+      cat("*****Could not unzip file: ", paste0(state, ".zip"), "\n")
+    }
   }
+
+  # connect to FIADB
+  con <- dbConnect(RSQLite::SQLite(), FIADBFile)
+
+  # read tables
+  PLOT <- dbReadTable(con, "PLOT")
+
+  # disconnect from database
+  dbDisconnect(con)
 }
-
-# connect to FIADB
-con <- dbConnect(RSQLite::SQLite(), FIADBFile)
-
-# read tables
-PLOT <- dbReadTable(con, "PLOT")
 totalPlots <- nrow(PLOT)
-
-# disconnect from database
-dbDisconnect(con)
 
 # get plots measured in last ~10 years...could include multiple measurements for some plots
 PLOT <- subset(PLOT, MEASYEAR >= 2010)
